Move CrossRepositoryBlobMountsSupport to .configuration
