+++ .gitignore
@@ -9,3 +9,4 @@ out
 /.settings
 /.classpath
 /.project
+jib-*/bin
