+++ jib-core/src/test/java/com/google/cloud/tools/jib/configuration/CrossRepositoryBlobMountsSupportTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.jib.builder;
+package com.google.cloud.tools.jib.configuration;
 
 import com.google.cloud.tools.jib.JibLogger;
 import com.google.cloud.tools.jib.configuration.BuildConfiguration;
